+++ src/index.ts
@@ -13,6 +13,7 @@ export enum ProcessingResult {
 export interface CallbackPayload {
   fieldValueType: string;
   fieldKey: string;
+  fieldValue: any;
 }
 export interface FunctionCallbackPayload {
   processingStrategy: ProcessingType;
@@ -30,29 +31,29 @@ export type OnErrorPayload = CallbackPayload & FunctionCallbackPayload & OnError
 export type OnSuccessPayload = CallbackPayload & FunctionCallbackPayload & OnSuccessAdditionalPayload;
 export type OnNonFunctionPayload = CallbackPayload;
 export type OnErrorResult = OnErrorAdditionalPayload["functionError"] | void;
-export type OnSuccessResult = void;
-export type OnNonFunctionResult = void;
+export type OnSuccessResult = OnSuccessPayload["functionResult"] | void;
+export type OnNonFunctionResult = CallbackPayload["fieldValue"] | void;
 export type OnError = (payload: OnErrorPayload) => OnErrorResult;
 export type OnSuccess = (payload: OnSuccessPayload) => OnSuccessResult;
 export type OnNonFunction = (payload: OnNonFunctionPayload) => OnNonFunctionResult;
 
-export type ProxyHandlerGenericExecutionOptions = {
+export type InterceptorOptions = {
   onError: OnError;
   onSuccess: OnSuccess;
   onNonFunction: OnNonFunction;
   /**
    * @description
-   *`callbackEnding` this field is used when an object (on which this proxy handler is applied) function use callback style to return value and have callback to promise transformation function. This field define a name of a callback to promise transformation function which should be executed after primary function execution.
+   *`callbackEnding` field is used when an object (on which proxy with returned handler is applied) function use callback style to return value and have callback to promise transformation function. This field define a name of a callback to promise transformation function which should be executed after primary function execution.
    * @example
    * ```
    * import { StepFunctions } from "aws-sdk";
-   * import { proxyHandlerGenericExecution } from "proxy-handler-generic-execution";
+   * import { interceptor } from "generic-interceptor";
    *
    * const callbackEnding = "promise";
    * const stepFunctions = new StepFunctions();
    * const wrappedStepFunctions = new Proxy(
    *   stepFunctions,
-   *   proxyHandlerGenericExecution({
+   *   interceptor({
    *     callbackEnding,
    *     onSuccess: () => {},
    *     onNonFunction: () => {},
@@ -68,20 +69,22 @@ export type ProxyHandlerGenericExecutionOptions = {
   callbackEnding?: CallbackEnding;
 };
 
-export type ProxyHandlerGenericExecution = (options: ProxyHandlerGenericExecutionOptions) => ProxyHandler<any>;
+export type Interceptor = (options: InterceptorOptions) => ProxyHandler<any>;
 
-export const proxyHandlerGenericExecution: ProxyHandlerGenericExecution = (options) => ({
+export const interceptor: Interceptor = (options) => ({
   get: (target, key) => {
-    const fieldValueType: CallbackPayload["fieldValueType"] = typeof target[key];
+    const fieldValue: OnNonFunctionPayload["fieldValue"] = target[key];
+    const fieldValueType: CallbackPayload["fieldValueType"] = typeof fieldValue;
+    const commonCallbackPayload: CallbackPayload = { fieldKey: String(key), fieldValueType, fieldValue };
     if (fieldValueType !== "function") {
-      options.onNonFunction({ fieldKey: String(key), fieldValueType });
-      return target[key] as NonFunctionFieldValue;
+      options.onNonFunction(commonCallbackPayload);
+      return fieldValue as NonFunctionFieldValue;
     }
 
     return (...functionArgs: FunctionCallbackPayload["functionArgs"]) => {
       try {
         const syncResult: unknown | Promise<unknown> | (unknown & { [key: string]: () => Promise<unknown> }) = (
-          target[key] as ExecutionFunction
+          fieldValue as ExecutionFunction
         )(...functionArgs);
         const callbackEnding = options.callbackEnding;
         if (
@@ -95,25 +98,21 @@ export const proxyHandlerGenericExecution: ProxyHandlerGenericExecution = (optio
             [callbackEnding]: async (...functionArgs2: FunctionCallbackPayload["functionArgs"]) => {
               try {
                 const asyncResult = await syncResult[callbackEnding](...functionArgs2);
-                options.onSuccess({
+                return (options.onSuccess({
+                  ...commonCallbackPayload,
                   functionArgs,
                   processingResult: ProcessingResult.succeed,
                   functionResult: asyncResult,
-                  fieldValueType,
-                  fieldKey: String(key),
                   processingStrategy: ProcessingType.callbackEnding,
-                });
-                return asyncResult;
+                }) || asyncResult) as OnSuccessPayload["functionResult"];
               } catch (error) {
-                error = (options.onError({
+                throw (options.onError({
+                  ...commonCallbackPayload,
                   functionArgs,
                   processingResult: ProcessingResult.failed,
                   functionError: error,
-                  fieldValueType,
-                  fieldKey: String(key),
                   processingStrategy: ProcessingType.callbackEnding,
                 }) || error) as OnErrorPayload["functionError"];
-                throw error;
               }
             },
           };
@@ -122,47 +121,39 @@ export const proxyHandlerGenericExecution: ProxyHandlerGenericExecution = (optio
           return (async () => {
             try {
               const asyncResult = await syncResult;
-              options.onSuccess({
+              return (options.onSuccess({
+                ...commonCallbackPayload,
                 functionArgs,
                 processingResult: ProcessingResult.succeed,
                 functionResult: asyncResult,
-                fieldValueType,
-                fieldKey: String(key),
                 processingStrategy: ProcessingType.promise,
-              });
-              return asyncResult;
+              }) || asyncResult) as OnSuccessPayload["functionResult"];
             } catch (error) {
-              error = (options.onError({
+              throw (options.onError({
+                ...commonCallbackPayload,
                 functionArgs,
                 processingResult: ProcessingResult.failed,
                 functionError: error,
-                fieldValueType,
-                fieldKey: String(key),
                 processingStrategy: ProcessingType.promise,
               }) || error) as OnErrorPayload["functionError"];
-              throw error;
             }
           })();
         }
-        options.onSuccess({
+        return (options.onSuccess({
+          ...commonCallbackPayload,
           functionArgs,
           processingResult: ProcessingResult.succeed,
           functionResult: syncResult,
-          fieldValueType,
-          fieldKey: String(key),
           processingStrategy: ProcessingType.synchronous,
-        });
-        return syncResult;
+        }) || syncResult) as OnSuccessPayload["functionResult"];
       } catch (error) {
-        error = (options.onError({
+        throw (options.onError({
+          ...commonCallbackPayload,
           functionArgs,
           processingResult: ProcessingResult.failed,
           functionError: error,
-          fieldValueType,
-          fieldKey: String(key),
           processingStrategy: ProcessingType.synchronous,
         }) || error) as OnErrorPayload["functionError"];
-        throw error;
       }
     };
   },