+++ tests/common/no-function-common.spec.ts
@@ -4,15 +4,19 @@ import { commonDescribe, commonIt } from "../shared/jest";
 
 commonDescribe("<common prefix> <no function common>", () => {
   commonIt("<common prefix> <get no functional property test name>", () => {
-    const strategyInput = { [parameterName]: "value" };
+    const fieldValue = "value";
+    const strategy = { [parameterName]: fieldValue };
     const dataset = prepareDataset({ options: { onNonFunction: jest.fn() } });
-    const proxy = prepareProxy(dataset, strategyInput);
+    const proxy = prepareProxy(strategy, dataset);
 
-    expect(proxy[parameterName]).toEqual(strategyInput[parameterName]);
+    const result = proxy[parameterName];
+
+    expect(result).toEqual(fieldValue);
     expect(dataset.options.onNonFunction).toHaveBeenCalledTimes(1);
     expect(dataset.options.onNonFunction).toHaveBeenCalledWith({
       fieldKey: parameterName,
-      fieldValueType: typeof proxy[parameterName],
+      fieldValue,
+      fieldValueType: typeof fieldValue,
     } as OnNonFunctionPayload);
   });
 });